Fix the resource GET result when running multiple profiles
The issue only happens when running multiple profiles
with RequiredProfiles.

Root cause: For different profiles, we merge their
results into one final result and for the same URI
the latter result would overwrite the previous one.

For the 'success' attribute which is updated in
validateSingleURI function, it only set to true when
both conditions met:
1. GET is success
2. Current profile have requirements on this resource

If a profile don't have requirements on this resource,
even the GET is success, the property would not be
updated. This patch fixed this issue by updating
that property once the first condition is met.

Signed-off-by: Jordan Chen <jordanchen@google.com>